+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr (development version)
 
+* `cur_data()` and `cur_data_all()` don't simplify list columns in rowwise data frames (#5901).
+
 * `storms` data updated to 2020 (@steveharoz, #5899).
 
 * `coalesce()` accepts 1-D arrays (#5557).